Space save

+++ src/cart/components/positioncomponent.nim
@@ -2,6 +2,6 @@ import ../ecs/ecs
 
 type
    PositionComponent* = ref object of Component
-      x*: int32
-      y*: int32
-      z*: int32
+      x*: uint8
+      y*: uint8
+      z*: uint8
